feat: Use one workspace for everything
