--- Example/Example-objc/SceneDelegate.h
@@ -1,16 +1,0 @@
-//
-//  SceneDelegate.h
-//  Example-objc
-//
-//  Created by Daniel Griesser on 03.03.20.
-//  Copyright © 2020 Sentry. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface SceneDelegate : UIResponder <UIWindowSceneDelegate>
-
-@property (strong, nonatomic) UIWindow * window;
-
-@end
-
