--- Example/Example-objc/main.m
@@ -1,19 +1,0 @@
-//
-//  main.m
-//  Example-objc
-//
-//  Created by Daniel Griesser on 03.03.20.
-//  Copyright © 2020 Sentry. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-#import "AppDelegate.h"
-
-int main(int argc, char * argv[]) {
-    NSString * appDelegateClassName;
-    @autoreleasepool {
-        // Setup code that might create autoreleased objects goes here.
-        appDelegateClassName = NSStringFromClass([AppDelegate class]);
-    }
-    return UIApplicationMain(argc, argv, nil, appDelegateClassName);
-}
