--- Example/Podfile
@@ -1,14 +1,0 @@
-#platform :ios, '8.0'
-use_frameworks!
-
-target 'Example' do
-    pod 'Sentry', :path => '../'
-end
-
-target 'Example-objc' do
-    pod 'Sentry', :path => '../'
-end
-
-target 'MacExampleSwift' do
-    pod 'Sentry', :path => '../'
-end
