--- Example/README.md
@@ -1,14 +1,0 @@
-<p align="center">
-    <a href="https://sentry.io" target="_blank" align="center">
-        <img src="https://sentry-brand.storage.googleapis.com/sentry-logo-black.png" width="280">
-    </a>
-<br/>
-    <h1>Example Swift iOS App using sentry-cocoa</h1>
-</p>
-
-# Usage
-
-- Make sure you have CocoaPods installed https://cocoapods.org/
-- Run `pod install` in this folder
-- Open `Example.xcworkspace` 
-- Open `AppDelegate.swift` and paste your DSN
