+++ Samples/iOS-ObjectiveC/iOS-ObjectiveC/AppDelegate.h
