+++ Samples/iOS-ObjectiveC/iOS-ObjectiveC/Assets.xcassets/Contents.json
