+++ Samples/iOS-ObjectiveC/iOS-ObjectiveC/SceneDelegate.h
@@ -1,0 +1,16 @@
+//
+//  SceneDelegate.h
+//  iOS-ObjectiveC
+//
+//  Created by Daniel Griesser on 06.04.20.
+//  Copyright © 2020 Sentry. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface SceneDelegate : UIResponder <UIWindowSceneDelegate>
+
+@property (strong, nonatomic) UIWindow * window;
+
+@end
+
