+++ Samples/iOS-Swift/iOS-Swift/AppDelegate.swift
