+++ Samples/iOS-Swift/iOS-Swift/Base.lproj/LaunchScreen.storyboard
