+++ Samples/iOS-Swift/iOS-Swift/Info.plist
