+++ Samples/macOS-Swift/macOS-Swift/Assets.xcassets/AppIcon.appiconset/Contents.json
