+++ Samples/macOS-Swift/macOS-Swift/Info.plist
