+++ Samples/macOS-Swift/macOS-Swift/macOS_Swift.entitlements
