+++ Sentry.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:Samples/macOS-Swift/macOS-Swift.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Samples/iOS-ObjectiveC/iOS-ObjectiveC.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Samples/iOS-Swift/iOS-Swift.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Sentry.xcodeproj">
+   </FileRef>
+</Workspace>
