+++ Sentry.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
